clipme(): remove some obsolete diagnostic code.
This #if completely replaced the actually useful version of clipme(),
so I think it must have been used early in development before the
useful version was even written.

Since then it's bit-rotted to the point where it doesn't even make
sense: the "#endif" is nested inside a while loop that the "#if 0" and
"#else" are outside, so that if the condition were changed to evaluate
true, you'd get syntactic nonsense out of the preprocessor.

And I can't see any use for it now, even if it did compile! Get rid of
it.