Fix Xterm216+ Alt-Fn on Windows
On most keyboard modes the escape-sequence for Alt-Fn is the same as
for Fn alone but with an additional ESC prepended.

However, this additional ESC should not be sent for keyboard modes
that sends a different escape-sequence when Alt is pressed, like e.g.
Xterm216+, as this is not expected by the server-side application.

Signed-off-by: Anders Larsen <al@alarsen.net>